deployment/requirements.txt
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
requests>=2.23.0
torch>=1.7.0
torchvision>=0.8.1